Move setup-request.js to /authorise directory
Remove app/setup-request/index.js
