+++ app/authorise/index.js
@@ -3,8 +3,10 @@ const { generateRedirectUri } = require('./authorise-uri');
 const { setTokenPayload, accessToken } = require('./access-tokens');
 const { authorisationCodeGrantedHandler } = require('./authorisation-code-granted');
 const { createAccessToken } = require('./obtain-access-token');
+const { accessTokenAndResourcePath } = require('./setup-request');
 
 exports.accessToken = accessToken;
+exports.accessTokenAndResourcePath = accessTokenAndResourcePath;
 exports.authorisationCodeGrantedHandler = authorisationCodeGrantedHandler;
 exports.createClaims = createClaims;
 exports.createJsonWebSignature = createJsonWebSignature;
