--- app/setup-request/index.js
@@ -1,3 +1,0 @@
-const { accessTokenAndResourcePath } = require('./setup-request');
-
-exports.accessTokenAndResourcePath = accessTokenAndResourcePath;
